slettet unødvendig fil, og la til en main utenfor pakke
