+++ frequence/frequence.go
@@ -1,0 +1,5 @@
+package frequence
+
+func test() {
+
+}
