+++ main_bfrequence.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/heiaviking/is-105/is105-ica03/frequence"
+
+func main() {
+	frequence.Bfrequence()
+}
